Minor portability tweaks.

+++ hdf/test/tutils.h
@@ -40,21 +40,21 @@ int         num_errs
 /* Use %ld to print the value because long could cover most cases. */
 /* Used to make certain a return value _is_not_ a value */
 #define CHECK(ret, val, where) \
-do {if (Verbosity>9) printf("   Call to HDF routine: %15s at line %4d in %s returned %ld \n",where,__LINE__,__FILE__,ret);\
-if(ret == val) {printf("*** UNEXPECTED RETURN from %s is %ld at line %4d in %s\n", where, ret, __LINE__,__FILE__); num_errs++;} \
+do {if (Verbosity>9) printf("   Call to HDF routine: %15s at line %4d in %s returned %ld \n",where,(int)__LINE__,__FILE__,(long)ret);\
+if(ret == val) {printf("*** UNEXPECTED RETURN from %s is %ld at line %4d in %s\n", where, (long)ret, (int)__LINE__,__FILE__); num_errs++;} \
 } while(0)
 
 /* Used to make certain a return value _is_ a value */
 #define VERIFY(x, val, where) \
-do {if (Verbosity>9) printf("   Call to HDF routine: %15s at line %4d in %s had value %ld \n",where,__LINE__,__FILE__,x);\
-if(x != val) {printf("*** UNEXPECTED VALUE from %s is %ld at line %4d in %s\n", where, x,__LINE__,__FILE__); num_errs++;} \
+do {if (Verbosity>9) printf("   Call to HDF routine: %15s at line %4d in %s had value %ld \n",where,(int)__LINE__,__FILE__,(long)x);\
+if(x != val) {printf("*** UNEXPECTED VALUE from %s is %ld at line %4d in %s\n", where, (long)x,(int)__LINE__,__FILE__); num_errs++;} \
 } while(0)
 
 #define RESULT(a) \
 do { \
-if (Verbosity>8) printf("   Call to HDF routine: %15s at line %4d in %s returned %ld \n",a,__LINE__,__FILE__,ret); \
+if (Verbosity>8) printf("   Call to HDF routine: %15s at line %4d in %s returned %ld \n",a,(int)__LINE__,__FILE__,(long)ret); \
 if (Verbosity>9) HEprint(stdout,0); \
-if(ret == FAIL) {printf("*** UNEXPECTED RETURN from %s is %ld at line %4d in %s\n", a, ret,__LINE__,__FILE__); num_errs++;} \
+if(ret == FAIL) {printf("*** UNEXPECTED RETURN from %s is %ld at line %4d in %s\n", a, (long)ret,(int)__LINE__,__FILE__); num_errs++;} \
 } while(0)
 
 #define MESSAGE(v,a) {if (Verbosity>v) {a}}
